+++ clodoo/bck_filestore.sh
@@ -30,7 +30,7 @@ fi
 TESTDIR=$(findpkg "" "$TDIR . .." "tests")
 RUNDIR=$(readlink -e $TESTDIR/..)
 
-__version__=0.3.35.2
+__version__=0.3.35.3
 
 
 OPTOPTS=(h        b          n            q           t       V           v)