+++ clodoo/junk/check_4_seq.sh
@@ -1,4 +1,4 @@
-# __version__=0.3.35.2
+# __version__=0.3.35.3
 THIS=$(basename $0)
 if [ -z "$1" ]; then
   echo "$0 db [user]"