+++ clodoo/junk/set_color.sh
@@ -16,7 +16,7 @@ if [ -z "$Z0LIBDIR" ]; then
   exit 2
 fi
 
-__version__=0.3.35.2
+__version__=0.3.35.3
 
 
 OPTOPTS=(h        n            V           v)