+++ python_plus/tests/python_plus_test_03.py
@@ -25,7 +25,7 @@ parser.add_argument('-V')
 ctx = parser.parseoptargs(['-V'])
 """
 
-__version__ = "1.0.3.2"
+__version__ = "1.0.3.3"
 
 
 def version():