+++ egg-info/history.rst
@@ -1,3 +1,15 @@
+python_plus: 1.0.3.3 (2021-09-26)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+[FIX] vem: error message "riga 99: deactivate:"
+
+
+clodoo: 0.3.35.3 (2021-09-26)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [FIX] clodoo: specific openpyxl for python2
+
+
 travis_emulator: 1.0.2.1 (2021-09-25)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -7,7 +19,8 @@ travis_emulator: 1.0.2.1 (2021-09-25)
 python_plus: 1.0.3.2 (2021-09-25)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-[FIX] vem: create ven with -I switch
+[FIX] vem: create vem with -I switch
+[REF] vem refactoring
 
 
 wok_code: 1.0.2.2.2 (2021-09-24)